FiveM Bait car Standalone
Any vehicle can be tagged as a Tragic Bait Car with an intuitive menu. Tragic Bait Cars can be remotely disabled by the player that installed the Tragic Bait Car system regardless of who is driving the vehicle. Players are unable to enter/exit a disabled Tragic Bait Car, and players inside the vehicle will be given a fake wanted level and an alert.